ID&P did request can now be skipped

+++ offering_catalog/src/main/java/it/nextworks/tmf_offering_catalog/services/ProductOfferingService.java
@@ -81,9 +81,9 @@ public class ProductOfferingService {
         }
     }
 
-    public ProductOffering create(ProductOfferingCreate productOfferingCreate)
+    public ProductOffering create(ProductOfferingCreate productOfferingCreate, Boolean skipIDP)
             throws IOException, DIDGenerationRequestException, StakeholderNotRegisteredException,
-            NotExistingEntityException, NullIdentifierException {
+            NotExistingEntityException, NullIdentifierException, DIDAlreadyRequestedForProductException, ScLcmRequestException {
 
         final String id = UUID.randomUUID().toString();
         log.info("Storing Product Offering " + id + ".");
@@ -145,16 +145,40 @@ public class ProductOfferingService {
         productOfferingRepository.save(productOffering);
         log.info("Product Offering " + id + " stored in Catalog.");
 
-        log.info("Requesting DID via CommunicationService to ID&P for Product Offering " + id + ".");
+        if(skipIDP != null) {
+            if(skipIDP) {
+                ProductOfferingStatus productOfferingStatus = new ProductOfferingStatus()
+                        .catalogId(id)
+                        .did(null)
+                        .status(ProductOfferingStatesEnum.DID_REQUESTED);
+                productOfferingStatusRepository.save(productOfferingStatus);
 
-        try {
-            communicationService.requestDID(id, ow.getToken());
-        } catch (DIDGenerationRequestException e) {
-            productOfferingRepository.delete(productOffering);
-            throw e;
+                communicationService.handleDIDReceiving(id, UUID.randomUUID().toString());
+            } else {
+                log.info("Requesting DID via CommunicationService to ID&P for Product Offering " + id + ".");
+
+                try {
+                    communicationService.requestDID(id, ow.getToken());
+                } catch (DIDGenerationRequestException e) {
+                    productOfferingRepository.delete(productOffering);
+                    throw e;
+                }
+
+                log.info("DID successfully requested via CommunicationService to ID&P.");
+            }
+        } else {
+            log.info("Requesting DID via CommunicationService to ID&P for Product Offering " + id + ".");
+
+            try {
+                communicationService.requestDID(id, ow.getToken());
+            } catch (DIDGenerationRequestException e) {
+                productOfferingRepository.delete(productOffering);
+                throw e;
+            }
+
+            log.info("DID successfully requested via CommunicationService to ID&P.");
         }
 
-        log.info("DID successfully requested via CommunicationService to ID&P.");
         log.info("Product Offering created with id " + id + ".");
 
         return productOffering;
